fix(export): prevent dynamic time units
We utilized the String() method of the time.Duration type when writing
results to an export file. In some cases this resulted in results being
exported in multiple duration units. A common case were sub-millisecond
results which the String() method turned into a microsecond results.
Nice for the CLI, bad for gnuplot as we told it that all results would
be in milliseconds. So a 0.7 millisecond result was turned into 700
microsecond result which led to gnuplot rendering the datapoint as
700ms, which obviously is completely wrong.

This fix forces all duration fields to be exported in milliseconds.